+++ packages/presets/src/theme/recipes/radio-group.ts
@@ -33,15 +33,30 @@ export const radioGroup = defineSlotRecipe({
           background: 'accent.default',
         },
       },
+      _disabled: {
+        borderColor: 'border.disabled',
+        color: 'fg.disabled',
+
+        _hover: {
+          bg: 'initial',
+          color: 'fg.disabled',
+        },
+      },
     },
     radio: {
       alignItems: 'center',
       cursor: 'pointer',
       display: 'flex',
+      _disabled: {
+        cursor: 'not-allowed',
+      },
     },
     radioLabel: {
       color: 'fg.emphasized',
       fontWeight: 'medium',
+      _disabled: {
+        color: 'fg.disabled',
+      },
     },
   },
   defaultVariants: {